Add paralell run tests which paralell runs with DataProvider
