+++ src/main/java/pages/OutBox.java
@@ -1,15 +1,20 @@
 package pages;
 
+import core.driver.WebDriverSingleton;
 import core.elements.impl.Button;
 import core.model.Message;
 import org.openqa.selenium.*;
 import org.openqa.selenium.support.FindBy;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.Wait;
+import org.openqa.selenium.support.ui.WebDriverWait;
+
 import java.util.List;
 
 public class OutBox extends BasePage{
     @FindBy(xpath = "//div[@class='TN bzz aHS-bnu']")
     private Button outBox;
-    @FindBy(xpath = "//table//tr[@class='zA yO']")
+    @FindBy(xpath = "//table//span[@class='bqe']")
     private WebElement subject;
 
     public OutBox(WebDriver dr) {
@@ -18,7 +23,15 @@ public class OutBox extends BasePage{
 
     public void goToOutBox(){
         outBox.click();
+        waitOutbox();
+
     }
+
+public void waitOutbox(){
+    Wait<WebDriver> wait = new WebDriverWait(WebDriverSingleton.getDriver(), 10).ignoring(StaleElementReferenceException.class, ElementNotVisibleException.class);
+    wait.until(ExpectedConditions.titleContains("Надіслані"));
+}
+
 
     public boolean checkMessageIsPresent(Message msg){
         String sub;
