+++ src/main/resources/Run CheckMailBox paralell.xml
@@ -1,0 +1,7 @@
+<suite name="CheckOutbox Suite" >
+    <test name="CheckOutbox">
+        <classes>
+            <class name="CheckMailOutbox" />
+        </classes>
+    </test>
+</suite>
