+++ src/main/resources/Run SendMail paralell.xml
@@ -1,0 +1,7 @@
+<suite name="SendMail Suite" >
+    <test name="SendMail">
+        <classes>
+            <class name="SendMail" />
+        </classes>
+    </test>
+</suite>
