+++ src/main/resources/RunTests.xml
@@ -1,4 +1,4 @@
-<suite name="Test-class Suite" parallel="classes" thread-count="3" >
+<suite name="Test-class Suite"  parallel="methods" thread-count="3">
     <test name="Test-class test" >
         <classes>
             <class name="CheckMailOutbox" />
