+++ system.properties
@@ -1,8 +1,8 @@
-#Fri Dec 21 11:47:19 EET 2018
-address=mail.google.com
+#Fri Dec 21 15:28:54 EET 2018
 login=test18.acount18
-site=https\://mail.google.com
 password=qwerty18()
-subject=67016582
+site=https\://mail.google.com
+subject=29692273
 sendTo=olga.dudar@gmail.com
 driverPath=\\drivers\\chromedriver.exe
+e.com=
